Refactors internals to named exports
