+++ src/v2/view-builder/internals/index.js
@@ -1,0 +1,17 @@
+import BaseHeader from './BaseHeader';
+import BaseFooter from './BaseFooter';
+import BaseForm from './BaseForm';
+import BaseFormWithPolling from './BaseFormWithPolling';
+import BaseModel from './BaseModel';
+import BaseView from './BaseView';
+import FormInputFactory from './FormInputFactory';
+
+export const { addCustomButton, createCustomButtons, createIdpButtons } = FormInputFactory;
+export {
+  BaseHeader,
+  BaseFooter,
+  BaseForm,
+  BaseFormWithPolling,
+  BaseModel,
+  BaseView,
+};
