+++ src/v2/view-builder/views/IdentifyRecoveryView.js
@@ -1,7 +1,5 @@
 import { loc } from 'okta';
-import BaseView from '../internals/BaseView';
-import BaseForm from '../internals/BaseForm';
-import BaseFooter from '../internals/BaseFooter';
+import { BaseForm, BaseFooter, BaseView } from '../internals';
 
 const Body = BaseForm.extend({
