+++ src/v2/view-builder/views/TerminalView.js
@@ -1,8 +1,5 @@
 import { createCallout, loc } from 'okta';
-import BaseView from '../internals/BaseView';
-import BaseForm from '../internals/BaseForm';
-import BaseFooter from '../internals/BaseFooter';
-import BaseHeader from '../internals/BaseHeader';
+import { BaseHeader, BaseForm, BaseFooter, BaseView } from '../internals';
 import HeaderBeacon from '../components/HeaderBeacon';
 import { getBackToSignInLink, getSkipSetupLink } from '../utils/LinksUtil';
 import { getIconClassNameForBeacon } from '../utils/AuthenticatorUtil';
