+++ src/v2/view-builder/views/consent/AdminConsentView.js
@@ -1,7 +1,6 @@
-import BaseView from '../../internals/BaseView';
+import { BaseView } from '../../internals';
 import AdminConsentViewHeader from './AdminConsentViewHeader';
 import ConsentViewForm from './ConsentViewForm';
-
 
 export default BaseView.extend({
   Header: AdminConsentViewHeader,
