+++ src/v2/view-builder/views/consent/ConsentViewForm.js
@@ -1,5 +1,5 @@
 import { loc } from 'okta';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm } from '../../internals';
 
 export default BaseForm.extend(
   {
