+++ src/v2/view-builder/views/device/SSOExtensionView.js
@@ -1,5 +1,5 @@
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm } from '../../internals';
 import Util from '../../../../util/Util';
 import { loc } from 'okta';
