+++ src/v2/view-builder/views/device/SignInDeviceView.js
@@ -1,6 +1,5 @@
 import { $, _, loc } from 'okta';
-import BaseView from '../../internals/BaseView';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm, BaseView } from '../../internals';
 import IdentifierFooter from '../../components/IdentifierFooter';
 import SignInWithDeviceOption from '../signin/SignInWithDeviceOption';
 import Link from '../../components/Link';
