+++ src/v2/view-builder/views/idp/AuthenticatorIdPView.js
@@ -1,5 +1,5 @@
 import { _, loc } from 'okta';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm } from '../../internals';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import AuthenticatorEnrollFooter from '../../components/AuthenticatorEnrollFooter';
 import AuthenticatorVerifyFooter from '../../components/AuthenticatorVerifyFooter';
