+++ src/v2/view-builder/views/password/ChallengeAuthenticatorPasswordView.js
@@ -1,5 +1,5 @@
 import { loc } from 'okta';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm } from '../../internals';
 import AuthenticatorVerifyFooter from '../../components/AuthenticatorVerifyFooter';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import { getForgotPasswordLink } from '../../utils/LinksUtil';
