+++ src/v2/view-builder/views/webauthn/EnrollWebauthnView.js
@@ -1,5 +1,5 @@
 import { _, loc, createCallout, createButton } from 'okta';
-import BaseForm from '../../internals/BaseForm';
+import { BaseForm } from '../../internals';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import webauthn from '../../../../util/webauthn';
 import CryptoUtil from '../../../../util/CryptoUtil';
